pylint: auto fixes

+++ helpers/dbtools.py
@@ -13,9 +13,8 @@ from pyrogram.errors import (
     PeerIdInvalid,
     UserIsBlocked,
 )
-from pyrogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 
-from config import BROADCAST_AS_COPY, GROUP_SUPPORT, UPDATES_CHANNEL, LOG_CHANNEL
+from config import BROADCAST_AS_COPY, GROUP_SUPPORT, LOG_CHANNEL
 from helpers.database import db, dcmdb
 
 
@@ -30,9 +29,8 @@ async def handle_user_status(bot, cmd):
 
     ban_status = await db.get_ban_status(chat_id)
     if ban_status["is_banned"]:
-        if (
-            datetime.date.today() - datetime.date.fromisoformat(ban_status["banned_on"])
-        ).days > ban_status["ban_duration"]:
+        if (datetime.date.today() - datetime.date.fromisoformat(
+                ban_status["banned_on"])).days > ban_status["ban_duration"]:
             await db.remove_ban(chat_id)
         else:
             await cmd.reply_text(
@@ -72,7 +70,9 @@ async def main_broadcast_handler(m, db):
     all_users = await db.get_all_users()
     broadcast_msg = m.reply_to_message
     while True:
-        broadcast_id = "".join(random.choice(string.ascii_letters) for i in range(3))
+        broadcast_id = "".join(
+            random.choice(
+                string.ascii_letters) for i in range(3))
         if not broadcast_ids.get(broadcast_id):
             break
     out = await m.reply_text(
